afl: for mmap at different offsets for each open instance.
There are some weird behaviour wen starting multiple instances of AFL when
we add an extra sync() in its run_target loop.  The symptons are that only
one instance can be started, the other complaining that there was no
instrumentation output (the afl area was read as all 0).

Try to work-around what ever is causing this by using different offsets
when mmap'ing /dev/afl in the different processes so that nobody can think
pages in the multiple afl-fuzz are the same (as they were all mmap()ing
with MAP_SHARED from offset zero).  A bit far fetched but heh, it doesn't
hurt anyway!

Signed-off-by: Quentin Casasnovas <quentin.casasnovas@oracle.com>