afl: change disassoc logic.
Following a change in afl-fuzz design, where it runs the wrapper in its own
context rather than using the fork server, we change the semantic for the
disassoc command, where before, any process with the file descriptor could
disassoc another task from the afl_area.  Now, it disassociates the current
task only.

The added bonus of that change is that we don't really need any lock on the
task_struct anymore since only the same process can disassociates, and that
simplified a lot the afl_maybe_log function which doesn't need to take any
locks anymore.

Signed-off-by: Quentin Casasnovas <quentin.casasnovas@oracle.com>